v2/backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.0
django-cors-headers>=4.3
django-redis>=5.4
django-jazzmin>=2.6
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
requests>=2.28
python-telegram-bot>=21.0
